+++ cmd/network_validator_deposit.go
@@ -12,6 +12,7 @@ import (
 )
 
 const CommandOptionGasPrice = "gasPrice"
+const CommandOptionDry = "dry"
 
 // depositCmd represents the deposit command
 var depositCmd = &cobra.Command{
@@ -23,10 +24,16 @@ address. Remember it will need your wallet address and private keys. Thus it wil
 This tool is necessary to activate new validators`,
 	Example: "lukso network validator deposit",
 	Run: func(cmd *cobra.Command, args []string) {
+		dry, _ := cmd.Flags().GetBool(CommandOptionDry)
+		if dry {
+			fmt.Println("THIS IS A DRY RUN")
+		}
 		nodeConf := network.MustGetNodeConfig()
+
 		valSecrets := nodeConf.GetValSecrets()
-		if valSecrets == nil {
-			cobra.CompErrorln("no validator credentials exist")
+		if valSecrets == nil || valSecrets.Deposit == nil || valSecrets.Eth1Data == nil {
+			utils.PrintColoredError("no validator credentials exist. Did you forget to setup your validators?")
+			utils.Coloredln("    lukso network validator setup")
 			return
 		}
 
@@ -35,13 +42,33 @@ This tool is necessary to activate new validators`,
 			cobra.CompErrorln(err.Error())
 			return
 		}
-		totalDeposits, err := network.Deposit(valSecrets.Deposit.DepositFileLocation, valSecrets.Deposit.ContractAddress, valSecrets.Eth1Data.WalletPrivKey, nodeConf.ApiEndpoints.ExecutionApi, gasPrice)
+
+		vc := nodeConf.ValidatorCredentials
+		// should never happen
+		if vc == nil {
+			cobra.CompErrorln("couldn't find contract details")
+			return
+		}
+
+		events, err := network.NewDepositEvents(vc.Deposit.ContractAddress, nodeConf.ApiEndpoints.ExecutionApi)
+		if err != nil {
+			cobra.CompErrorln(fmt.Sprintf("couldn't load deposit data from contract, reason: %s", err.Error()))
+			return
+		}
+
+		fmt.Println("Past deposit events loaded", len(events.Events))
+
+		totalDeposits, err := network.Deposit(&events, valSecrets.Deposit.DepositFileLocation, valSecrets.Deposit.ContractAddress, valSecrets.Eth1Data.WalletPrivKey, nodeConf.ApiEndpoints.ExecutionApi, gasPrice, dry)
 		if err != nil {
 			cobra.CompErrorln(err.Error())
 			return
 		}
-		fmt.Printf("You successfully deposited %d key(s)! Your keys need to be activated which takes around 8h. You can check the status by calling:\n", totalDeposits)
-		fmt.Println(utils.ConsoleInBlue("        lukso network validator describe"))
+		if !dry {
+			fmt.Printf("You successfully deposited %d key(s)! Your keys need to be activated which takes around 8h. You can check the status by calling:\n", totalDeposits)
+			fmt.Println(utils.ConsoleInBlue("        lukso network validator describe"))
+		} else {
+			fmt.Println("THIS WAS A DRY RUN - you didn't deposit any keys")
+		}
 	},
 }
 
@@ -49,4 +76,5 @@ func init() {
 	validatorCmd.AddCommand(depositCmd)
 
 	depositCmd.Flags().Int64P(CommandOptionGasPrice, "g", 1000000, "set the gas price for transactions")
+	depositCmd.Flags().BoolP(CommandOptionDry, "d", false, "don't run the transactions but just prepare it")
 }