+++ cmd/network_validator_describe.go
@@ -26,26 +26,40 @@ var validatorDescribeCmd = &cobra.Command{
 	Example: "lukso network validator describe",
 	Run: func(cmd *cobra.Command, args []string) {
 		key, _ := cmd.Flags().GetString("key")
-		baseUrl, _ := cmd.Flags().GetString("beaconapi")
-		if baseUrl == "" {
-			baseUrl = network.GetDefaultNodeConfigByOptionParam(viper.GetString(network.CommandOptionChain)).ApiEndpoints.ConsensusApi
-		}
 		epoch, _ := cmd.Flags().GetInt64("epoch")
 
+		nodeConf := network.LoadNodeConfOrDefault(network.GetChainByString(viper.GetString(network.CommandOptionChain)))
+		baseUrl := nodeConf.ApiEndpoints.ConsensusApi
+		vc := nodeConf.ValidatorCredentials
+
+		// should never happen
+		if vc == nil {
+			cobra.CompErrorln("couldn't find contract details")
+			return
+		}
+
+		events, err := network.NewDepositEvents(vc.Deposit.ContractAddress, nodeConf.ApiEndpoints.ExecutionApi)
+		if err != nil {
+			cobra.CompErrorln(fmt.Sprintf("couldn't load deposit data from contract, reason: %s", err.Error()))
+			return
+		}
+
+		// key was provided in command option
 		if key != "" {
-			err := describe([]string{key}, baseUrl, epoch)
+			err := describe(&events, []string{key}, baseUrl, epoch)
 			if err != nil {
 				cobra.CompErrorln(err.Error())
 			}
 			return
 		}
 
-		nodeConf := network.MustGetNodeConfig()
-
-		baseUrl = nodeConf.ApiEndpoints.ConsensusApi
-
 		valSecrets := nodeConf.GetValSecrets()
 		if valSecrets == nil {
+			cobra.CompErrorln(network.ErrMsgValidatorSecretNotPresent)
+			return
+		}
+
+		if valSecrets.Eth1Data == nil {
 			cobra.CompErrorln(network.ErrMsgValidatorSecretNotPresent)
 			return
 		}
@@ -77,7 +91,7 @@ var validatorDescribeCmd = &cobra.Command{
 			pubKeys[k] = d.PubKey
 		}
 
-		err = describe(pubKeys, baseUrl, epoch)
+		err = describe(&events, pubKeys, baseUrl, epoch)
 		if err != nil {
 			cobra.CompErrorln(err.Error())
 		}
@@ -85,21 +99,29 @@ var validatorDescribeCmd = &cobra.Command{
 }
 
 func maybeAddHexPrefix(address string) string {
-	if strings.Contains(address, "0x") {
-		return address
+	a := address
+	if !strings.Contains(address, "0x") {
+		a = fmt.Sprintf("0x%s", address)
 	}
-
-	return fmt.Sprintf("0x%s", address)
+	return strings.ToLower(a)
 }
 
-func describe(pubKeys []string, baseUrl string, epoch int64) error {
+func describe(events *network.DepositEvents, pubKeys []string, baseUrl string, epoch int64) error {
 	beaconClient := beaconapi.NewBeaconClient(baseUrl)
 	for _, d := range pubKeys {
 		pubKey := maybeAddHexPrefix(d)
-		fmt.Println("....................................................................................")
+		amount := events.Amount(pubKey)
+		if amount == 0 {
+			fmt.Printf("Validator %s has no deposits yet...\n", pubKey)
+			// nothing to do anymore, validator doesn't have a deposit yet
+			//scontinue
+		}
+		fmt.Println("........................................................................................................................................................................")
+		utils.ColoredPrintln("Validator", pubKey)
+		utils.ColoredPrintln("Staked Amount", amount)
 		state, status, err := beaconClient.ValidatorState(pubKey, epoch)
 		if status == http.StatusNotFound {
-			utils.PrintColoredError(fmt.Sprintf("validator %s not found on consensus node", pubKey))
+			fmt.Printf("Validator %s is pending...\n", pubKey)
 			continue
 		}
 		if err != nil {
@@ -122,7 +144,9 @@ func describe(pubKeys []string, baseUrl string, epoch int64) error {
 
 func init() {
 	validatorCmd.AddCommand(validatorDescribeCmd)
-	validatorDescribeCmd.Flags().StringP("beaconapi", "b", "", "endpoint of beacon api")
 	validatorDescribeCmd.Flags().Int64P("epoch", "e", -1, "epoch to be described - if left out it is the head epoch")
 	validatorDescribeCmd.Flags().StringP("key", "k", "", "validator key to be described - keep empty to describe your validators defined in keystore")
 }
+
+// /eth/v1/beacon/states/head/validators/0xa066a1e4bc483eb1471d21fd24907513c049edf78837df4e391174d96c46f354799723de06f5aa814ba5787b485a8d75
+// /eth/v1/beacon/states/head/validators/0x4c2f1829606061722f5bde51ceb5c482d90b37dda0fcc0146148a5925811e3c9