+++ src/utils/console.go
@@ -128,6 +128,10 @@ func ColoredPrintln(label string, value interface{}) {
 	fmt.Print(ConsoleWriteLabeledValueI(label, value))
 }
 
+func Coloredln(msg string) {
+	fmt.Println(ConsoleInColor(ConsoleColorBrightBlue, msg))
+}
+
 func PrintColoredError(msg string) {
 	fmt.Println(ConsoleInColor(ConsoleColorRed, fmt.Sprintf("error: %s", msg)))
 }